Trace QImage::Format as int in QImageData::create()
Traces can contain only a few types; for now, enums have to be converted
to int. Amends 6d323c0b221a3b127d80e338a42306cc34bc4d40

Task-number: QTBUG-102862

+++ src/gui/image/qimage.cpp
@@ -99,7 +99,7 @@ QImageData * QImageData::create(const QSize &size, QImage::Format format)
     if (size.isEmpty() || format <= QImage::Format_Invalid || format >= QImage::NImageFormats)
         return nullptr;                             // invalid parameter(s)
 
-    Q_TRACE_SCOPE(QImageData_create, size, format);
+    Q_TRACE_SCOPE(QImageData_create, size, static_cast<int>(format));
 
     int width = size.width();
     int height = size.height();
